% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_graph.R
\name{as_pipeline_graph}
\alias{as_pipeline_graph}
\title{Convert pipeline objects or igraph to a tidygraph tbl_graph}
\usage{
as_pipeline_graph(x, ...)
}
\arguments{
\item{x}{A list of FilePath/FileUses/FileOutputs objects, or an igraph object.}

\item{...}{Additional arguments passed to tidygraph::as_tbl_graph.}
}
\value{
A tbl_graph object.
}
\description{
Enables native dplyr verbs (activate, filter, mutate, group_by, summarise)
on computational graphs.
}
