% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_graph.R
\name{as_tbl_graph.list}
\alias{as_tbl_graph.list}
\title{Convert pipeline list to tidygraph}
\usage{
\method{as_tbl_graph}{list}(x, ...)
}
\arguments{
\item{x}{A list of FilePath/FileUses/FileOutputs objects.}

\item{...}{Additional arguments passed to \code{\link{as_pipeline_graph}}.}
}
\value{
A tbl_graph object.
}
\description{
A list of pipeline objects is converted by [as_pipeline_graph()]; any other list is passed
on to tidygraph's own list method, which this method would otherwise replace.
}
