% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbe_database_relationships.R
\name{autoplot.cbe_key_integrity}
\alias{autoplot.cbe_key_integrity}
\title{Autoplot Method for Key Integrity}
\usage{
\method{autoplot}{cbe_key_integrity}(object, ...)
}
\arguments{
\item{object}{A \code{\link{cbe_check_key_integrity}} object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A ggplot2 object.
}
\description{
Autoplot Method for Key Integrity
}
