% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_model.R
\name{autoplot.cv_joint_model}
\alias{autoplot.cv_joint_model}
\title{Autoplot Method for Cross-Validated Joint Models}
\usage{
\method{autoplot}{cv_joint_model}(object, ...)
}
\arguments{
\item{object}{A \code{cv_joint_model} object.}

\item{...}{Additional arguments.}
}
\value{
A ggplot2 object.
}
\description{
Visualizes comparative predictive performance across paradigms,
plotting the distribution of Integrated Brier Scores and Concordance.
}
