% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sas_script.R
\name{cbe_sas_macro_dir}
\alias{cbe_sas_macro_dir}
\alias{sas_macro_dir}
\title{Directory of TempleCBE SAS Macros}
\usage{
cbe_sas_macro_dir()

sas_macro_dir()
}
\value{
Absolute path to the SAS macro directory.
}
\description{
Returns the file system path to the directory containing TempleCBE's
bundled SAS macros (`cbe_brier_score.sas`, `cbe_cox_phreg.sas`,
`cbe_counting_process.sas`, and `cbe_macros.sas`).
}
\examples{
try(cbe_sas_macro_dir())
}
\seealso{
[cbe_sas_macro_path()], [run_sas_script()], [find_sas()]
}
