% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_graph.R
\name{controls_datasets}
\alias{controls_datasets}
\title{Check if a FileOutputs object controls datasets (outputs an rds, xlsx, etc. file)}
\usage{
controls_datasets(obj)
}
\arguments{
\item{obj}{A FileOutputs object.}
}
\value{
Logical.
}
\description{
Check if a FileOutputs object controls datasets (outputs an rds, xlsx, etc. file)
}
