% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_graph.R
\docType{class}
\name{FileOutputs-class}
\alias{FileOutputs-class}
\alias{show,FileOutputs-method}
\title{FileOutputs S4 Class}
\usage{
\S4method{show}{FileOutputs}(object)
}
\arguments{
\item{object}{A \code{\linkS4class{FileOutputs}} object.}
}
\description{
Represents an executable or rendering pipeline file that produces downstream output artifacts.
}
\details{
Inherits from \code{\linkS4class{FileUses}} (inheriting upstream \code{dependencies}),
while adding a list of downstream output \code{\linkS4class{FilePath}} objects produced when
this file executes or renders.
}
\section{Slots}{

\describe{
\item{\code{output}}{List of \code{\linkS4class{FilePath}} objects produced by this file.}
}}

\seealso{
\code{\linkS4class{FilePath}}, \code{\linkS4class{FileUses}}, \code{\link{FileOutputs}}, \code{\link{create_qmd_renderer}}
}
