% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_graph.R
\docType{class}
\name{FileUses-class}
\alias{FileUses-class}
\alias{show,FileUses-method}
\title{FileUses S4 Class}
\usage{
\S4method{show}{FileUses}(object)
}
\arguments{
\item{object}{A \code{\linkS4class{FileUses}} object.}
}
\description{
Represents a pipeline file that depends on other upstream files.
Extends \code{\linkS4class{FilePath}} by adding upstream input dependency tracking.
}
\details{
A \code{FileUses} node represents a computational artifact whose execution or validity
relies upon one or more input \code{\linkS4class{FilePath}} objects listed in \code{dependencies}.
}
\section{Slots}{

\describe{
\item{\code{dependencies}}{List of \code{\linkS4class{FilePath}} objects that this file depends on.}
}}

\seealso{
\code{\linkS4class{FilePath}}, \code{\linkS4class{FileOutputs}}, \code{\link{FileUses}}
}
