% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_graph.R
\name{FileUses}
\alias{FileUses}
\title{Construct a FileUses object}
\usage{
FileUses(
  name,
  path,
  dependencies = list(),
  renders = FALSE,
  stage = NA_character_,
  artifact_role = NA_character_,
  description = NA_character_
)
}
\arguments{
\item{name}{Character. User-supplied human-readable name for the file.}

\item{path}{Character. Absolute or relative file system path.}

\item{dependencies}{List of \code{\linkS4class{FilePath}} objects that this file depends on.}

\item{renders}{Logical. Indicates whether the file is a rendered file or rendering script. Defaults to \code{FALSE}.}

\item{stage}{Character. Pipeline lifecycle stage.}

\item{artifact_role}{Character. Pipeline role of the artifact.}

\item{description}{Character. Brief description of the file.}
}
\value{
A new \code{\linkS4class{FileUses}} object.
}
\description{
User-friendly constructor function for instantiating a \code{\linkS4class{FileUses}} object.
}
