% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plug_and_play_schema.R
\name{find_section_file}
\alias{find_section_file}
\title{Find Section Data File in a Directory}
\usage{
find_section_file(data_dir, index)
}
\arguments{
\item{data_dir}{Directory to search.}

\item{index}{Section name/token to look for.}
}
\value{
File path.
}
\description{
Find Section Data File in a Directory
}
