% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cbe.R
\name{fmt_num}
\alias{fmt_num}
\title{Format Numbers to Fixed Decimals}
\usage{
fmt_num(x, digits = 2)
}
\arguments{
\item{x}{Numeric value}

\item{digits}{Number of decimal places (default: 2)}
}
\value{
Formatted character string
}
\description{
Format Numbers to Fixed Decimals
}
