% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cbe.R
\name{fmt_pct}
\alias{fmt_pct}
\title{Format a Percentage}
\usage{
fmt_pct(x, digits = 1)
}
\arguments{
\item{x}{Numeric proportion between 0 and 1}

\item{digits}{Number of decimal places (default: 1)}
}
\value{
Formatted character string with '\%'
}
\description{
Format a Percentage
}
