% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cbe.R
\name{fmt_sig}
\alias{fmt_sig}
\title{Format Significant Figures}
\usage{
fmt_sig(x, digits = 3)
}
\arguments{
\item{x}{Numeric value}

\item{digits}{Number of significant figures (default: 3)}
}
\value{
Formatted character string
}
\description{
Format Significant Figures
}
