% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infix_helpers.R
\name{infix_helpers}
\alias{infix_helpers}
\alias{\%notin\%}
\alias{\%like\%}
\alias{\%ilike\%}
\alias{\%flike\%}
\title{Pattern Matching and Infix Operator Utilities}
\usage{
x \%notin\% table

x \%like\% pattern

x \%ilike\% pattern

x \%flike\% pattern
}
\arguments{
\item{x}{Vector to test.}

\item{table}{Vector of values to test against (for \code{\%notin\%}).}

\item{pattern}{Pattern to match (for \code{\%like\%}, \code{\%flike\%}, \code{\%ilike\%}).}
}
\value{
A logical vector.
}
\description{
Useful infix operators for string matching and logical negation.
}
