% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_utils.R
\name{install.packages.no_lock}
\alias{install.packages.no_lock}
\title{Install a Package, Bypassing an Existing Lock}
\usage{
install.packages.no_lock(packages)
}
\arguments{
\item{packages}{Character vector of package name(s) to install.}
}
\value{
Invisibly, \code{NULL} (as per \code{\link[utils]{install.packages}}).
}
\description{
Install a Package, Bypassing an Existing Lock
}
\examples{
\dontrun{
install.packages.no_lock("dplyr")
}
}
