% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_test.R
\name{is.int}
\alias{is.int}
\title{Is a Vector Composed of Integer-Valued Numbers}
\usage{
is.int(col)
}
\arguments{
\item{col}{A numeric vector.}
}
\value{
A single logical.
}
\description{
Is a Vector Composed of Integer-Valued Numbers
}
\examples{
is.int(sample(-100:100, size = 500, replace = TRUE))
is.int(runif(500))
}
