% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_plots.R
\name{pca_feature_loading_heatmap}
\alias{pca_feature_loading_heatmap}
\title{PCA Feature-Loading Heatmap}
\usage{
pca_feature_loading_heatmap(pca_model)
}
\arguments{
\item{pca_model}{A \code{\link[stats]{prcomp}} object.}
}
\value{
A ggplot object.
}
\description{
Heatmap of each original feature's loading onto each principal component.
}
\examples{
pca_model <- prcomp(mtcars, center = TRUE, scale. = TRUE)
pca_feature_loading_heatmap(pca_model)
}
