% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cox.R
\name{plot_cox}
\alias{plot_cox}
\title{Cox Proportional Hazards Diagnostic and Reporting Visualizations}
\description{
Functions for creating hazard ratio forest plots, model-predicted survival curves,
and continuous predictor vs. event probability diagnostic plots.
}
\keyword{internal}
