% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_plots.R
\name{plot_pca_bi}
\alias{plot_pca_bi}
\title{PCA Biplot}
\usage{
plot_pca_bi(pca_model, newdata, column, x = 1, y = 2)
}
\arguments{
\item{pca_model}{A \code{\link[stats]{prcomp}} object.}

\item{newdata}{Data to project onto \code{pca_model}.}

\item{column}{Column in \code{newdata} to use as point labels.}

\item{x, y}{Which principal components to plot on the x/y axes (default 1, 2).}
}
\value{
A ggplot object.
}
\description{
Biplot of feature loading vectors against two principal components.
}
\examples{
pca_model <- prcomp(mtcars, center = TRUE, scale. = TRUE)
mtcars2 <- tibble::rownames_to_column(mtcars, "model")
plot_pca_bi(pca_model, mtcars2, column = "model")
}
