% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbe_cox_single.R
\name{print.cbe_cox}
\alias{print.cbe_cox}
\title{Print Method for cbe_cox Object}
\usage{
\method{print}{cbe_cox}(x, ...)
}
\arguments{
\item{x}{Object of class \code{cbe_cox}}

\item{...}{Additional arguments (unused)}
}
\description{
Print Method for cbe_cox Object
}
