% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbe_cox_check.R
\name{print.cbe_cox_check}
\alias{print.cbe_cox_check}
\title{Print Method for cbe_cox_check Object}
\usage{
\method{print}{cbe_cox_check}(x, ...)
}
\arguments{
\item{x}{Object of class \code{cbe_cox_check}}

\item{...}{Additional arguments (unused)}
}
\description{
Print Method for cbe_cox_check Object
}
