% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbe_cox_multi.R
\name{print.cbe_cox_multi}
\alias{print.cbe_cox_multi}
\title{Print Method for cbe_cox_multi Object}
\usage{
\method{print}{cbe_cox_multi}(x, ...)
}
\arguments{
\item{x}{Object of class \code{cbe_cox_multi}}

\item{...}{Additional arguments (unused)}
}
\description{
Print Method for cbe_cox_multi Object
}
