% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbe_km_single.R
\name{print.cbe_km}
\alias{print.cbe_km}
\title{Print Method for cbe_km Object}
\usage{
\method{print}{cbe_km}(x, ...)
}
\arguments{
\item{x}{Object of class \code{cbe_km}}

\item{...}{Additional arguments (unused)}
}
\description{
Print Method for cbe_km Object
}
