% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_model.R
\name{print.joint_model}
\alias{print.joint_model}
\title{Print Method for Joint Models}
\usage{
\method{print}{joint_model}(x, ...)
}
\arguments{
\item{x}{A \code{joint_model} object.}

\item{...}{Additional arguments.}
}
\description{
Print Method for Joint Models
}
