% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cbe.R
\name{scale_color_cbe}
\alias{scale_color_cbe}
\title{Discrete Color Scale for Temple CBE}
\usage{
scale_color_cbe(...)
}
\arguments{
\item{...}{Arguments passed to \code{ggplot2::scale_color_manual}}
}
\description{
Discrete Color Scale for Temple CBE
}
