% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cbe.R
\name{scale_fill_cbe}
\alias{scale_fill_cbe}
\title{Discrete Fill Scale for Temple CBE}
\usage{
scale_fill_cbe(...)
}
\arguments{
\item{...}{Arguments passed to \code{ggplot2::scale_fill_manual}}
}
\description{
Discrete Fill Scale for Temple CBE
}
