% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_error.R
\name{sd.error}
\alias{sd.error}
\title{Compute standard error}
\usage{
sd.error(x, na.rm = FALSE)
}
\arguments{
\item{x}{vector of data}

\item{na.rm}{should NA values be removed, TRUE or FALSE, default is FALSE}
}
\value{
numeric value representing standard error
}
\description{
Compute standard error
}
\examples{
sd.error(mtcars$mpg)

}
