% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_test.R
\name{significance_stars}
\alias{significance_stars}
\title{P-value Significance Stars}
\usage{
significance_stars(p_value)
}
\arguments{
\item{p_value}{A numeric vector of p-values.}
}
\value{
A character vector of significance stars.
}
\description{
P-value Significance Stars
}
\examples{
significance_stars(c(0.0001, 0.02, 0.2, 0.8))
}
