% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_graph.R
\name{summary.tbl_graph}
\alias{summary.tbl_graph}
\title{S3 summary method for tbl_graph objects}
\usage{
\method{summary}{tbl_graph}(object, ...)
}
\description{
S3 summary method for tbl_graph objects
}
