% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_model.R
\name{tidy.joint_model}
\alias{tidy.joint_model}
\title{Tidy Method for Joint Models}
\usage{
\method{tidy}{joint_model}(x, ...)
}
\arguments{
\item{x}{A \code{joint_model} object.}

\item{...}{Additional arguments.}
}
\value{
A tibble with comparative coefficients per feature.
}
\description{
Extracts and aligns nonzero coefficients across the survival (\code{coxnet}),
binary classification (\code{status}), and duration regression (\code{time}) models.
}
