% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cbe.R
\name{words}
\alias{words}
\title{Combine Character Vector into Natural-Language Words}
\usage{
words(x, none = "none")
}
\arguments{
\item{x}{Character vector}

\item{none}{Fallback text when vector is empty (default: "none")}
}
\value{
Character string (e.g., "A, B, and C")
}
\description{
Combine Character Vector into Natural-Language Words
}
