% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_graph.R
\name{create_qmd_renderer}
\alias{create_qmd_renderer}
\title{Create a FileOutputs object for a rendering QMD}
\usage{
create_qmd_renderer(
  name,
  path,
  deps = list(),
  file_stage = NA_character_,
  output_format = "pdf",
  description = NA_character_
)
}
\arguments{
\item{name}{The nickname for the file.}

\item{path}{The full path to the .qmd file.}

\item{deps}{A list of dependency objects (e.g., list(og_DATA)).}

\item{file_stage}{The 'stage' for this file.}

\item{output_format}{"pdf" or "html". Default is "pdf".}

\item{description}{A brief summary of the document.}
}
\value{
A FileOutputs object.
}
\description{
Create a FileOutputs object for a rendering QMD
}
