% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_percent_miss.R,
%   R/plot.features_percent_miss.R
\name{features_percent_miss}
\alias{features_percent_miss}
\alias{plot.features_percent_miss}
\title{Calculate Percentage of Missing Data Per Feature}
\usage{
features_percent_miss(data)

\method{plot}{features_percent_miss}(x, ...)
}
\arguments{
\item{data}{A data frame or tibble.}

\item{x}{An object of class 'features_percent_miss' returned by the
\code{\link{features_percent_miss}} function.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A tibble with columns `feature`, `SumNa`, `SumComp`, `PctNa`, `PctComp` sorted descending by `PctNa`.

A ggplot object representing the bar plot.
}
\description{
Computes the count and percentage of missing (`NA`) and complete values for each column in a dataset.

This function creates a bar plot showing the percentage of complete values
for each feature in a data frame or tibble.
}
\examples{
df <- data.frame(a = c(1, 2, NA, 4), b = c(NA, NA, 3, 4))
features_percent_miss(df)
res <- features_percent_miss(mtcars)
plot(res)

}
