% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_outliers.R
\name{flag_outliers}
\alias{flag_outliers}
\title{Flag and Classify Outliers}
\usage{
flag_outliers(col)
}
\arguments{
\item{col}{A numeric vector or column.}
}
\value{
A one-column-input-turned-tibble with \code{value}, \code{.outlier}
  (factor: \code{FALSE}, \code{TRUE}) and \code{.outlier_type} (factor: \code{"NONE"}, \code{"MILD"},
  or \code{"EXTREME"}).
}
\description{
Flag and Classify Outliers
}
\examples{
flag_outliers(c(1, 2, 3, 4, 5, 100))
}
