% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_graph.R
\name{get_render_plan}
\alias{get_render_plan}
\title{Get the topologically sorted list of QMDs that need to be re-rendered}
\usage{
get_render_plan(all_objects)
}
\arguments{
\item{all_objects}{The full list of FilePath, FileUses, and FileOutputs objects.}
}
\value{
A list of FileOutputs objects (the QMDs) in the correct render order.
}
\description{
Get the topologically sorted list of QMDs that need to be re-rendered
}
