% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan_volcano_plot.R
\name{manhattan_plot}
\alias{manhattan_plot}
\title{Manhattan Plot}
\usage{
manhattan_plot(.data, var, log_p, alpha = 0.05, highlight_significant = TRUE)
}
\arguments{
\item{.data}{A data frame, e.g. the output of \code{\link{multiple_t_test}}.}

\item{var}{Unquoted column of variable labels.}

\item{log_p}{Unquoted column of \code{-log10(p.value)}.}

\item{alpha}{Significance threshold for the reference line/shading
(default \code{0.05}).}

\item{highlight_significant}{Logical (default \code{TRUE}); shade the
region above the \code{p < alpha} line.}
}
\value{
A ggplot object.
}
\description{
Manhattan Plot
}
\examples{
mtcars |>
  dplyr::mutate(am = factor(am)) |>
  multiple_t_test(.class = "am") |>
  manhattan_plot(var = var, log_p = log_p)
}
