% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_summary_table.R
\name{my_summary_table}
\alias{my_summary_table}
\title{Summary Table Function}
\usage{
my_summary_table(
  data,
  var,
  na.rm = FALSE,
  data.output = TRUE,
  table.output = FALSE
)
}
\arguments{
\item{data}{a dataframe or tibble}

\item{var}{variable to summarize}

\item{na.rm}{should NA values be removed, TRUE or FALSE, default is FALSE}

\item{data.output}{should the data be output, TRUE or FALSE, default is TRUE}

\item{table.output}{should the table be output, TRUE or FALSE, default is FALSE}
}
\value{
either a dataframe or readable table
}
\description{
Summary Table Function
}
\examples{
mtcars |> 
  dplyr::group_by(cyl) |> 
  my_summary_table(mpg)
  
mtcars |> 
  dplyr::group_by(cyl) |> 
  my_summary_table(mpg, table.output = TRUE)

}
