% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_pca.R
\name{proc_pca}
\alias{proc_pca}
\title{Process and Plot Principal Component Analysis (PCA)}
\usage{
proc_pca(pca_obj, data = NULL)
}
\arguments{
\item{pca_obj}{A `prcomp` object.}

\item{data}{Data frame used for PCA.}
}
\value{
Summary statistics or ggplot biplot object.
}
\description{
Helper functions for extracting component variances, feature loadings, and generating biplots.
}
