% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_famd.R
\name{step_famd}
\alias{step_famd}
\title{Factor Analysis of Mixed Data (FAMD) Recipe Step}
\usage{
step_famd(
  recipe,
  ...,
  role = "predictor",
  trained = FALSE,
  num_comp = 2,
  threshold = NA,
  options = list(),
  res = NULL,
  columns = NULL,
  prefix = "PC",
  keep_original_cols = FALSE,
  skip = FALSE,
  id = recipes::rand_id("famd")
)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of steps for this recipe.}

\item{...}{One or more selector functions to choose variables.}

\item{role}{Role for created variables (default "predictor").}

\item{trained}{Logical indicating if step has been trained.}

\item{num_comp}{Number of components to extract (default 2).}

\item{threshold}{Cumulative percentage of variance threshold.}

\item{options}{Options passed to underlying FAMD algorithm.}

\item{res}{FAMD object.}

\item{columns}{Vector of column names.}

\item{prefix}{Component prefix (default "PC").}

\item{keep_original_cols}{Logical; whether to retain original columns.}

\item{skip}{Logical; skip step when baking.}

\item{id}{Unique step identifier.}
}
\value{
An updated recipe object.
}
\description{
`step_famd` creates a *specification* of a recipe step that will extract
Factor Analysis of Mixed Data (FAMD) principal components from numeric and categorical variables.
}
