% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SumNa.R
\name{SumNa}
\alias{SumNa}
\title{Count Total Missing (NA) Values}
\usage{
SumNa(x)
}
\arguments{
\item{x}{A vector, matrix, or data frame.}
}
\value{
An integer representing the total count of missing values.
}
\description{
Calculates the total number of `NA` values across a vector, matrix, or data frame.
}
\examples{
SumNa(c(1, 2, NA, 4, NA))
SumNa(data.frame(a = c(1, NA), b = c(NA, 2)))
}
