% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan_volcano_plot.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Volcano Plot}
\usage{
volcano_plot(.data, log2_fold_change, log_p, var, highlight_significant = TRUE)
}
\arguments{
\item{.data}{A data frame, e.g. the output of \code{\link{multiple_t_test}}.}

\item{log2_fold_change}{Unquoted column of log2 fold-change.}

\item{log_p}{Unquoted column of \code{-log10(p.value)}.}

\item{var}{Unquoted column of variable labels.}

\item{highlight_significant}{Logical (default \code{TRUE}); shade the
region above the p < 0.05 line.}
}
\value{
A ggplot object.
}
\description{
Volcano Plot
}
\examples{
mtcars |>
  dplyr::mutate(am = factor(am)) |>
  multiple_t_test(.class = "am") |>
  volcano_plot(log2_fold_change = log2_fold_change, log_p = log_p, var = var)
}
