% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_norm.R
\name{z_norm}
\alias{z_norm}
\title{Z-Score Standard Normalization}
\usage{
z_norm(x, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector, matrix, or data frame.}

\item{na.rm}{Logical; whether to ignore NA values (default TRUE).}
}
\value{
Z-score standardized numeric object.
}
\description{
Standardizes numeric features to have mean = 0 and standard deviation = 1.
}
\examples{
z_norm(c(10, 20, 30, 40, 50))
}
